smb2: extract the pooled-session layer into fa_libsmb2_pool.{c,h}
fa_libsmb2.c grew from 1182 to 1687 lines with the connection pool, keepalive,
and URL/credential plumbing all inlined. The pool is a cohesive subsystem
(session struct, pool globals, 8 lifecycle functions, keepalive callout, and
the full credential-resolution + connect path) with a single well-defined
boundary: movian_smb2_session_acquire/release/invalidate and the parsed target.

Move that entire unit into fa_libsmb2_pool.c with a small fa_libsmb2_pool.h
interface. fa_libsmb2.c drops to 983 lines and now holds only the fa_protocol
VFS callbacks (scan/open/read/write/stat/...) plus the errno_to_fap helper it
still needs. No behavior change; the GDB write smoke and pool/echo smokes pass
unchanged (22 pool reuses vs 2 handshakes on a TrueNAS browse).
