+++ src/fileaccess/smb2/fa_libsmb2.c
@@ -7,6 +7,15 @@
  *  (at your option) any later version.
  */
 
+/*
+ * SMB2 fileaccess VFS backend.
+ *
+ * Implements the fa_protocol callbacks (scan/open/read/write/.../stat/...) on
+ * top of a pooled libsmb2 session provided by fa_libsmb2_pool.c. Each VFS op
+ * acquires a session, takes session->lock around its libsmb2 call(s) (libsmb2 is
+ * not thread safe and a context is bound to one socket), and releases when done.
+ */
+
 #include <errno.h>
 #include <fcntl.h>
 #include <inttypes.h>
@@ -22,81 +31,10 @@
 #include <smb2/libsmb2-dcerpc-srvsvc.h>
 
 #include "main.h"
-#include "keyring.h"
 #include "fileaccess/fa_proto.h"
-#include "misc/callout.h"
-#include "misc/queue.h"
 #include "misc/rstr.h"
 
-#define SMB2_DEFAULT_TIMEOUT 15
-
-/*
- * Connection pool limits.
- *
- * libsmb2 is not thread safe and binds a context to a single TCP socket, so a
- * session cannot be shared between threads without serialization. We keep one
- * long lived session per (server, share, user, domain) key in a pool so the
- * TCP+negotiate+session-setup+tree-connect handshake happens once and is reused
- * across the many VFS calls Movian issues while browsing a share. The cifs
- * native backend achieves the same with cifs_connections/cc_refcount.
- *
- * Idle sessions are kept around for a while so that bouncing back into a share
- * does not pay the handshake cost again, up to SMB2_POOL_MAX_SESSIONS. Older
- * sessions are reaped first.
- */
-#define SMB2_POOL_MAX_SESSIONS 8
-#define SMB2_POOL_IDLE_TTL_SEC 60
-
-/*
- * Keepalive cadence (seconds) and how many consecutive missed echoes are
- * tolerated before the session is torn down. Mirrors the cifs native
- * SMB_ECHO_INTERVAL / cc_wait_for_ping scheme.
- */
-#define SMB2_ECHO_INTERVAL 30
-#define SMB2_ECHO_MAX_MISSED 2
-
-#define SMB2TRACE(x, ...) do {                                 \
-    if(gconf.enable_smb_debug)                                 \
-      tracelog(0, TRACE_DEBUG, "SMB2-CLIENT", x, ##__VA_ARGS__); \
-  } while(0)
-
-typedef struct {
-  char *server;
-  char *share;
-  char *path;
-  char *user;
-  char *domain;
-} movian_smb2_target_t;
-
-typedef struct {
-  char *user;
-  char *password;
-  char *domain;
-} movian_smb2_credentials_t;
-
-typedef struct movian_smb2_session {
-  char *key;
-  char *server;
-  char *share;
-  char *user;
-  char *domain;
-  struct smb2_context *smb2;
-  hts_mutex_t lock;
-  int refcount;
-  int broken;
-  int64_t last_used;
-  callout_t keepalive;
-  int echo_pending;
-  int echo_missed;
-  LIST_ENTRY(movian_smb2_session) link;
-} movian_smb2_session_t;
-
-static LIST_HEAD(, movian_smb2_session) smb2_sessions =
-  LIST_HEAD_INITIALIZER(smb2_sessions);
-static hts_mutex_t smb2_pool_mutex;
-static int smb2_pool_inited;
-
-static void movian_smb2_keepalive_cb(callout_t *c, void *opaque);
+#include "fa_libsmb2_pool.h"
 
 typedef struct {
   fa_handle_t h;
@@ -111,187 +49,7 @@ typedef struct {
   struct srvsvc_NetrShareEnum_rep *reply;
 } movian_smb2_enum_state_t;
 
-
-static void
-movian_smb2_target_fini(movian_smb2_target_t *target)
-{
-  free(target->server);
-  free(target->share);
-  free(target->path);
-  free(target->user);
-  free(target->domain);
-}
-
-
-static int
-movian_smb2_target_parse(const char *url, movian_smb2_target_t *target,
-                         char *errbuf, size_t errlen)
-{
-  if(strncmp(url, "smb2://", 7)) {
-    snprintf(errbuf, errlen, "Invalid SMB2 URL");
-    return -1;
-  }
-
-  struct smb2_context *smb2 = smb2_init_context();
-  if(smb2 == NULL) {
-    snprintf(errbuf, errlen, "Unable to initialize SMB2");
-    return -1;
-  }
-
-  size_t parser_url_len = strlen(url);
-  char *parser_url = malloc(parser_url_len);
-  if(parser_url == NULL) {
-    snprintf(errbuf, errlen, "Out of memory while parsing SMB2 URL");
-    smb2_destroy_context(smb2);
-    return -1;
-  }
-  snprintf(parser_url, parser_url_len, "smb://%s", url + 7);
-
-  struct smb2_url *parsed = smb2_parse_url(smb2, parser_url);
-  free(parser_url);
-  if(parsed == NULL || parsed->server == NULL || parsed->server[0] == '\0') {
-    snprintf(errbuf, errlen, "Invalid SMB2 URL: %s", smb2_get_error(smb2));
-    if(parsed != NULL)
-      smb2_destroy_url(parsed);
-    smb2_destroy_context(smb2);
-    return -1;
-  }
-
-  int has_share = parsed->share != NULL;
-  int has_user = parsed->user != NULL;
-  int has_domain = parsed->domain != NULL;
-
-  target->server = strdup(parsed->server);
-  target->share = has_share ? strdup(parsed->share) : NULL;
-  target->path = parsed->path != NULL ? strdup(parsed->path) : strdup("");
-  target->user = has_user ? strdup(parsed->user) : NULL;
-  target->domain = has_domain ? strdup(parsed->domain) : NULL;
-
-  smb2_destroy_url(parsed);
-  smb2_destroy_context(smb2);
-
-  if(target->server == NULL || target->path == NULL ||
-     (has_share && target->share == NULL) ||
-     (has_user && target->user == NULL) ||
-     (has_domain && target->domain == NULL)) {
-    snprintf(errbuf, errlen, "Out of memory while parsing SMB2 URL");
-    movian_smb2_target_fini(target);
-    memset(target, 0, sizeof(*target));
-    return -1;
-  }
-  return 0;
-}
-
-
-static void
-movian_smb2_credentials_fini(movian_smb2_credentials_t *credentials)
-{
-  free(credentials->user);
-  free(credentials->password);
-  free(credentials->domain);
-}
-
-
-static char *
-movian_smb2_keyring_id(const movian_smb2_target_t *target)
-{
-  size_t len = strlen(target->server) + 32;
-
-  if(target->user != NULL)
-    len += strlen(target->user);
-  if(target->domain != NULL)
-    len += strlen(target->domain);
-
-  char *id = malloc(len);
-  if(id == NULL)
-    return NULL;
-
-  if(target->user != NULL) {
-    snprintf(id, len, "smb2:connection:%s%s%s@%s",
-             target->domain != NULL ? target->domain : "",
-             target->domain != NULL ? ";" : "",
-             target->user, target->server);
-  } else {
-    snprintf(id, len, "smb2:connection:%s", target->server);
-  }
-  return id;
-}
-
-
-static void
-movian_smb2_default_credentials(const movian_smb2_target_t *target,
-                                movian_smb2_credentials_t *credentials)
-{
-  credentials->user = strdup(target->user != NULL ? target->user : "guest");
-  credentials->password = NULL;
-  credentials->domain =
-    strdup(target->domain != NULL ? target->domain : "WORKGROUP");
-}
-
-
-static int
-movian_smb2_apply_target_identity(const movian_smb2_target_t *target,
-                                  movian_smb2_credentials_t *credentials,
-                                  char *errbuf, size_t errlen)
-{
-  if(target->user != NULL) {
-    char *user = strdup(target->user);
-    if(user == NULL) {
-      snprintf(errbuf, errlen, "Out of memory while preparing SMB2 login");
-      return -1;
-    }
-    free(credentials->user);
-    credentials->user = user;
-  }
-
-  if(target->domain != NULL) {
-    char *domain = strdup(target->domain);
-    if(domain == NULL) {
-      snprintf(errbuf, errlen, "Out of memory while preparing SMB2 login");
-      return -1;
-    }
-    free(credentials->domain);
-    credentials->domain = domain;
-  }
-  return 0;
-}
-
-
-static int
-movian_smb2_is_auth_error(int status, const char *reason)
-{
-  if(status == -EACCES || status == -EPERM)
-    return 1;
-  if(reason == NULL)
-    return 0;
-  return strstr(reason, "STATUS_LOGON_FAILURE") != NULL ||
-    strstr(reason, "STATUS_ACCESS_DENIED") != NULL ||
-    strstr(reason, "STATUS_NETWORK_ACCESS_DENIED") != NULL ||
-    strstr(reason, "STATUS_INVALID_PARAMETER") != NULL ||
-    strstr(reason, "STATUS_INVALID_ACCOUNT_NAME") != NULL ||
-    strstr(reason, "STATUS_WRONG_PASSWORD") != NULL ||
-    strstr(reason, "STATUS_ACCOUNT_RESTRICTION") != NULL ||
-    strstr(reason, "STATUS_INVALID_LOGON_HOURS") != NULL ||
-    strstr(reason, "STATUS_PASSWORD_EXPIRED") != NULL ||
-    strstr(reason, "STATUS_PASSWORD_MUST_CHANGE") != NULL ||
-    strstr(reason, "STATUS_ACCOUNT_DISABLED") != NULL ||
-    strstr(reason, "STATUS_ACCOUNT_EXPIRED") != NULL ||
-    strstr(reason, "STATUS_ACCOUNT_LOCKED_OUT") != NULL ||
-    strstr(reason, "STATUS_LOGON_NOT_GRANTED") != NULL ||
-    strstr(reason, "STATUS_LOGON_TYPE_NOT_GRANTED") != NULL;
-}
-
-
-static int
-movian_smb2_can_query_user(int flags)
-{
-  if(flags & (FA_NON_INTERACTIVE | FA_DISABLE_AUTH))
-    return 0;
-
-  char name[64];
-  return strcmp(hts_thread_name(name, sizeof(name)), "asyncio");
-}
-
+/* Error-code translation kept local to the VFS layer. */
 
 static fa_err_code_t
 movian_smb2_errno_to_fap(int status)
@@ -314,468 +72,6 @@ movian_smb2_errno_to_fap(int status)
   default:
     return FAP_ERROR;
   }
-}
-
-
-static int
-movian_smb2_session_key(const movian_smb2_target_t *target, const char *share,
-                        char **keyp)
-{
-  const char *user = target->user != NULL ? target->user : "";
-  const char *domain = target->domain != NULL ? target->domain : "";
-  const char *sh = share != NULL ? share : "";
-
-  size_t len = strlen(target->server) + strlen(sh) + strlen(user) +
-    strlen(domain) + 4;
-  char *key = malloc(len);
-  if(key == NULL)
-    return -1;
-
-  snprintf(key, len, "%s/%s/%s/%s", target->server, sh, user, domain);
-  *keyp = key;
-  return 0;
-}
-
-
-static void
-movian_smb2_session_destroy(movian_smb2_session_t *session)
-{
-  /*
-   * Caller must hold neither the pool nor the session lock. The session is
-   * already unlinked from the pool.
-   */
-  callout_disarm(&session->keepalive);
-  if(session->smb2 != NULL) {
-    smb2_disconnect_share(session->smb2);
-    smb2_destroy_context(session->smb2);
-  }
-  hts_mutex_destroy(&session->lock);
-  free(session->key);
-  free(session->server);
-  free(session->share);
-  free(session->user);
-  free(session->domain);
-  free(session);
-}
-
-
-/*
- * Reap idle and broken sessions. Called under the pool mutex. The least
- * recently used idle session is dropped when the pool grows past the limit.
- */
-static void
-movian_smb2_pool_evict_locked(void)
-{
-  if(LIST_EMPTY(&smb2_sessions))
-    return;
-
-  int64_t now = arch_get_ts();
-  int64_t cutoff = now - SMB2_POOL_IDLE_TTL_SEC * 1000000LL;
-
-  movian_smb2_session_t *session = LIST_FIRST(&smb2_sessions);
-  while(session != NULL) {
-    movian_smb2_session_t *next = LIST_NEXT(session, link);
-    if(session->refcount <= 0 &&
-       (session->broken || session->last_used < cutoff)) {
-      LIST_REMOVE(session, link);
-      movian_smb2_session_destroy(session);
-    }
-    session = next;
-  }
-
-  int count = 0;
-  LIST_FOREACH(session, &smb2_sessions, link)
-    count++;
-
-  while(count > SMB2_POOL_MAX_SESSIONS) {
-    movian_smb2_session_t *oldest = NULL;
-    LIST_FOREACH(session, &smb2_sessions, link) {
-      if(session->refcount > 0)
-        continue;
-      if(oldest == NULL || session->last_used < oldest->last_used)
-        oldest = session;
-    }
-    if(oldest == NULL)
-      break;
-    LIST_REMOVE(oldest, link);
-    movian_smb2_session_destroy(oldest);
-    count--;
-  }
-}
-
-
-static struct smb2_context *
-movian_smb2_connect_once(const movian_smb2_target_t *target,
-                         const char *share,
-                         const movian_smb2_credentials_t *credentials,
-                         int timeout, int *status,
-                         char *errbuf, size_t errlen)
-{
-  const char *user =
-    credentials->user != NULL && credentials->user[0] != '\0' ?
-    credentials->user : "guest";
-  struct smb2_context *smb2 = smb2_init_context();
-  if(smb2 == NULL) {
-    *status = -ENOMEM;
-    snprintf(errbuf, errlen, "Unable to initialize SMB2");
-    SMB2TRACE("Unable to initialize SMB2 context");
-    return NULL;
-  }
-
-  SMB2TRACE("Connecting to %s/%s timeout=%d",
-            target->server, share != NULL ? share : "", timeout);
-  SMB2TRACE("SETUP %s:%s:%s", user,
-            credentials->password != NULL ? "<set>" : "<unset>",
-            credentials->domain != NULL ? credentials->domain : "<unset>");
-
-  smb2_set_timeout(smb2, timeout);
-  smb2_set_security_mode(smb2, SMB2_NEGOTIATE_SIGNING_ENABLED);
-  smb2_set_user(smb2, user);
-  smb2_set_password(smb2, credentials->password);
-  if(credentials->domain != NULL)
-    smb2_set_domain(smb2, credentials->domain);
-
-  *status = smb2_connect_share(smb2, target->server, share, user);
-  if(*status == 0) {
-    SMB2TRACE("%s/%s Session setup", target->server,
-              share != NULL ? share : "");
-    return smb2;
-  }
-
-  snprintf(errbuf, errlen, "%s", smb2_get_error(smb2));
-  SMB2TRACE("SETUP status=%d reason=%s", *status, errbuf);
-  smb2_destroy_context(smb2);
-  return NULL;
-}
-
-
-/*
- * Resolve credentials and perform the (interactive, possibly retrying) connect.
- * Runs without the pool lock so that a keyring prompt does not stall other
- * sessions. Returns a fresh connected context and, on success, leaves the
- * resolved credentials in *resolved for the session to retain.
- */
-static struct smb2_context *
-movian_smb2_connect_impl(const movian_smb2_target_t *target, const char *share,
-                         int flags, int timeout, int *auth_needed,
-                         movian_smb2_credentials_t *resolved,
-                         char *errbuf, size_t errlen)
-{
-  movian_smb2_credentials_t credentials = {};
-  char *keyring_id = movian_smb2_keyring_id(target);
-  int status;
-
-  if(auth_needed != NULL)
-    *auth_needed = 0;
-
-  if(keyring_id == NULL) {
-    snprintf(errbuf, errlen, "Out of memory while preparing SMB2 login");
-    return NULL;
-  }
-
-  int keyring_status =
-    keyring_lookup(keyring_id, &credentials.user, &credentials.password,
-                   &credentials.domain, NULL, NULL, NULL, 0);
-  int have_saved_credentials = keyring_status == KEYRING_OK;
-  SMB2TRACE("Keyring lookup %s for %s",
-            have_saved_credentials ? "hit" : "miss", keyring_id);
-  if(keyring_status != KEYRING_OK) {
-    movian_smb2_default_credentials(target, &credentials);
-  } else if(movian_smb2_apply_target_identity(target, &credentials,
-                                               errbuf, errlen)) {
-    movian_smb2_credentials_fini(&credentials);
-    free(keyring_id);
-    return NULL;
-  }
-
-  char reason[256] = {};
-  struct smb2_context *smb2 =
-    movian_smb2_connect_once(target, share, &credentials, timeout, &status,
-                            reason, sizeof(reason));
-  if(smb2 != NULL) {
-    if(resolved != NULL) {
-      resolved->user = strdup(credentials.user != NULL ? credentials.user : "");
-      resolved->password =
-        strdup(credentials.password != NULL ? credentials.password : "");
-      resolved->domain =
-        strdup(credentials.domain != NULL ? credentials.domain : "");
-    }
-    movian_smb2_credentials_fini(&credentials);
-    free(keyring_id);
-    return smb2;
-  }
-
-  int auth_error = movian_smb2_is_auth_error(status, reason);
-  if(auth_error && auth_needed != NULL)
-    *auth_needed = 1;
-
-  int can_query_user = movian_smb2_can_query_user(flags);
-  char thread_name[64];
-  SMB2TRACE("Auth retry auth_error=%d can_query=%d flags=0x%x "
-            "thread=%s status=%d reason=%s",
-            auth_error, can_query_user, flags,
-            hts_thread_name(thread_name, sizeof(thread_name)), status,
-            reason);
-
-  if(!can_query_user || !auth_error) {
-    snprintf(errbuf, errlen, "Unable to connect to SMB2 share: %s", reason);
-    movian_smb2_credentials_fini(&credentials);
-    free(keyring_id);
-    return NULL;
-  }
-
-  movian_smb2_credentials_fini(&credentials);
-  memset(&credentials, 0, sizeof(credentials));
-  if(movian_smb2_apply_target_identity(target, &credentials,
-                                       errbuf, errlen)) {
-    free(keyring_id);
-    return NULL;
-  }
-  if(credentials.domain == NULL) {
-    credentials.domain = strdup("WORKGROUP");
-    if(credentials.domain == NULL) {
-      snprintf(errbuf, errlen, "Out of memory while preparing SMB2 login");
-      free(keyring_id);
-      return NULL;
-    }
-  }
-
-  size_t source_len = strlen(target->server) + 16;
-  char *source = malloc(source_len);
-  if(source == NULL) {
-    snprintf(errbuf, errlen, "Out of memory while preparing SMB2 login");
-    free(keyring_id);
-    return NULL;
-  }
-  snprintf(source, source_len, "SMB2 server '%s'", target->server);
-
-  keyring_status =
-    keyring_lookup(keyring_id,
-                   target->user != NULL ? NULL : &credentials.user,
-                   &credentials.password,
-                   target->domain != NULL ? NULL : &credentials.domain,
-                   NULL, source,
-                   have_saved_credentials ? reason : "Login required",
-                   KEYRING_QUERY_USER | KEYRING_SHOW_REMEMBER_ME |
-                   KEYRING_REMEMBER_ME_SET);
-  free(source);
-
-  if(keyring_status != KEYRING_OK) {
-    SMB2TRACE("Credential query rejected status=%d", keyring_status);
-    snprintf(errbuf, errlen, "Authentication rejected by user");
-    movian_smb2_credentials_fini(&credentials);
-    free(keyring_id);
-    return NULL;
-  }
-  SMB2TRACE("Credential query accepted");
-
-  smb2 = movian_smb2_connect_once(target, share, &credentials, timeout,
-                                  &status, reason, sizeof(reason));
-  if(smb2 == NULL)
-    snprintf(errbuf, errlen, "Unable to connect to SMB2 share: %s", reason);
-  else if(resolved != NULL) {
-    resolved->user = strdup(credentials.user != NULL ? credentials.user : "");
-    resolved->password =
-      strdup(credentials.password != NULL ? credentials.password : "");
-    resolved->domain =
-      strdup(credentials.domain != NULL ? credentials.domain : "");
-  }
-
-  movian_smb2_credentials_fini(&credentials);
-  free(keyring_id);
-  return smb2;
-}
-
-
-static void
-movian_smb2_pool_init_once(void)
-{
-  if(smb2_pool_inited)
-    return;
-  hts_mutex_init(&smb2_pool_mutex);
-  smb2_pool_inited = 1;
-}
-
-
-/*
- * Periodic keepalive. Runs on the main callout thread. Sends an SMB2 echo and,
- * if the previous echo never came back, counts it as a miss. After too many
- * misses (or a hard error) the session is marked broken so the next acquisition
- * reconnects. The echo itself is a synchronous libsmb2 call; it is short and
- * only ever sent for idle, refcount==0 sessions, so it never contends with a
- * streaming reader that already holds session->lock for its own data PDU.
- */
-static void
-movian_smb2_keepalive_cb(callout_t *c, void *opaque)
-{
-  movian_smb2_session_t *session = opaque;
-
-  hts_mutex_lock(&smb2_pool_mutex);
-  int still_alive = !session->broken && session->refcount <= 0;
-  if(session->echo_pending)
-    session->echo_missed++;
-  if(session->echo_missed > SMB2_ECHO_MAX_MISSED) {
-    SMB2TRACE("Keepalive %s/%s giving up (missed=%d) -> broken",
-              session->server, session->share, session->echo_missed);
-    session->broken = 1;
-  }
-  int broken = session->broken;
-  hts_mutex_unlock(&smb2_pool_mutex);
-
-  if(!broken && still_alive) {
-    hts_mutex_lock(&session->lock);
-    session->echo_pending = 1;
-    int rc = smb2_echo(session->smb2);
-    int missed = session->echo_missed;
-    if(rc == 0)
-      missed = session->echo_missed = 0;
-    session->echo_pending = 0;
-    hts_mutex_unlock(&session->lock);
-    SMB2TRACE("Keepalive %s/%s echo rc=%d missed=%d",
-              session->server, session->share, rc, missed);
-  }
-
-  if(!broken) {
-    hts_mutex_lock(&smb2_pool_mutex);
-    if(!session->broken)
-      callout_arm(&session->keepalive, movian_smb2_keepalive_cb, session,
-                  SMB2_ECHO_INTERVAL);
-    hts_mutex_unlock(&smb2_pool_mutex);
-  }
-}
-
-
-/*
- * Acquire a pooled session for (target, share). On a cache miss the TCP +
- * negotiate + session setup + tree connect handshake runs once; subsequent
- * acquisitions reuse the live session. Returns a session with refcount bumped
- * (locked for the caller via session->lock semantics — see release) or NULL on
- * failure.
- *
- * The returned session has its pool refcount incremented; the caller must pair
- * this with movian_smb2_session_release(). Per-operation serialization on a
- * single context is the caller's responsibility: take session->lock around the
- * libsmb2 call.
- */
-static movian_smb2_session_t *
-movian_smb2_session_acquire(const movian_smb2_target_t *target,
-                            const char *share, int flags, int timeout,
-                            int *auth_needed, char *errbuf, size_t errlen)
-{
-  movian_smb2_pool_init_once();
-
-  char *key;
-  if(movian_smb2_session_key(target, share, &key)) {
-    snprintf(errbuf, errlen, "Out of memory while opening SMB2 session");
-    return NULL;
-  }
-
-  movian_smb2_session_t *session = NULL;
-
-  hts_mutex_lock(&smb2_pool_mutex);
-  LIST_FOREACH(session, &smb2_sessions, link) {
-    if(!strcmp(session->key, key) && !session->broken) {
-      session->refcount++;
-      session->last_used = arch_get_ts();
-      hts_mutex_unlock(&smb2_pool_mutex);
-      free(key);
-      SMB2TRACE("Pool reuse %s/%s refcount=%d",
-                target->server, share != NULL ? share : "", session->refcount);
-      return session;
-    }
-  }
-  hts_mutex_unlock(&smb2_pool_mutex);
-
-  movian_smb2_credentials_t resolved = {};
-  struct smb2_context *smb2 =
-    movian_smb2_connect_impl(target, share, flags, timeout, auth_needed,
-                             &resolved, errbuf, errlen);
-  if(smb2 == NULL) {
-    free(key);
-    return NULL;
-  }
-
-  session = calloc(1, sizeof(*session));
-  if(session == NULL) {
-    snprintf(errbuf, errlen, "Out of memory while opening SMB2 session");
-    smb2_disconnect_share(smb2);
-    smb2_destroy_context(smb2);
-    movian_smb2_credentials_fini(&resolved);
-    free(key);
-    return NULL;
-  }
-
-  session->key = key;
-  session->smb2 = smb2;
-  session->refcount = 1;
-  session->last_used = arch_get_ts();
-  session->server = strdup(target->server);
-  session->share = strdup(share != NULL ? share : "");
-  session->user = resolved.user;
-  session->domain = resolved.domain;
-  if(session->server == NULL || session->share == NULL) {
-    snprintf(errbuf, errlen, "Out of memory while opening SMB2 session");
-    movian_smb2_credentials_fini(&resolved);
-    movian_smb2_session_destroy(session);
-    return NULL;
-  }
-  /*
-   * resolved.password is intentionally not retained: the session never
-   * re-authenticates (it is dropped when the server tears it down), and the
-   * keyring still holds the credentials for the next acquisition.
-   */
-  free(resolved.password);
-  resolved = (movian_smb2_credentials_t){};
-  hts_mutex_init(&session->lock);
-
-  hts_mutex_lock(&smb2_pool_mutex);
-  LIST_INSERT_HEAD(&smb2_sessions, session, link);
-  movian_smb2_pool_evict_locked();
-  callout_arm(&session->keepalive, movian_smb2_keepalive_cb, session,
-              SMB2_ECHO_INTERVAL);
-  hts_mutex_unlock(&smb2_pool_mutex);
-
-  SMB2TRACE("Pool create %s/%s refcount=1",
-            target->server, share != NULL ? share : "");
-  return session;
-}
-
-
-static void
-movian_smb2_session_release(movian_smb2_session_t *session)
-{
-  if(session == NULL)
-    return;
-
-  hts_mutex_lock(&smb2_pool_mutex);
-  session->refcount--;
-  session->last_used = arch_get_ts();
-  SMB2TRACE("Pool release %s/%s refcount=%d",
-            session->server, session->share, session->refcount);
-  if(session->refcount <= 0)
-    movian_smb2_pool_evict_locked();
-  hts_mutex_unlock(&smb2_pool_mutex);
-}
-
-
-/*
- * Mark a session as broken under the pool lock and drop it from the pool so the
- * next acquisition reconnects. Called when an operation fails in a way that
- * implies the context is no longer usable.
- */
-static void
-movian_smb2_session_invalidate(movian_smb2_session_t *session)
-{
-  if(session == NULL)
-    return;
-
-  hts_mutex_lock(&smb2_pool_mutex);
-  session->broken = 1;
-  if(session->refcount <= 0) {
-    LIST_REMOVE(session, link);
-    movian_smb2_session_destroy(session);
-  }
-  hts_mutex_unlock(&smb2_pool_mutex);
 }
